Tweak json and resolve minor mixup.
