+++ lib/src/symbolicate/v6/looked_up_addresses.rs
@@ -1,11 +1,12 @@
 use super::super::demangle;
-use crate::shared::{AddressDebugInfo, SymbolicationResult};
+use crate::shared::{AddressDebugInfo, InlineStackFrame, SymbolicationResult};
 use std::collections::HashMap;
 use std::ops::Deref;
 
 pub struct AddressResult {
-    pub symbol_name: String,
     pub symbol_address: u32,
+    pub symbol_name: String,
+    pub inline_frames: Option<Vec<InlineStackFrame>>,
 }
 
 pub struct LookedUpAddresses {
@@ -37,21 +38,22 @@ impl SymbolicationResult for LookedUpAddresses {
                     AddressResult {
                         symbol_address,
                         symbol_name,
+                        inline_frames: None,
                     },
                 )
             })
             .collect();
-        LookedUpAddresses {
-            address_results,
-            symbol_count,
-        }
+        LookedUpAddresses { address_results, symbol_count }
     }
 
     fn wants_address_debug_info() -> bool {
-        false
+        true
     }
 
-    fn add_address_debug_info(&mut self, _address: u32, _info: AddressDebugInfo) {
-        panic!("Should not be called")
+    fn add_address_debug_info(&mut self, address: u32, info: AddressDebugInfo) {
+        self.address_results
+            .get_mut(&address)
+            .unwrap()
+            .inline_frames = Some(info.frames);
     }
 }
